refactor: exclude stores and actions in app module exports
