+++ packages/desktop/views/dashboard/settings/views/general/CrashReporting.svelte
@@ -1,7 +1,7 @@
 <script lang="ts">
-    import { Checkbox, Text, Icon } from '@ui'
-    import { appSettings, initAppSettings } from '@core/app'
+    import { appSettings, initAppSettings } from '@core/app/stores'
     import { localize } from '@core/i18n'
+    import { Checkbox, Icon, Text } from '@ui'
 </script>
 
 <Text type="h4" classes="mb-3">{localize('views.settings.crashReporting.title')}</Text>
