+++ packages/desktop/views/dashboard/settings/views/general/Language.svelte
@@ -1,6 +1,6 @@
 <script lang="ts">
     import { Dropdown, Text } from '@ui'
-    import { appSettings } from '@core/app'
+    import { appSettings } from '@core/app/stores'
     import { SUPPORTED_LOCALES, localize, setLanguage } from '@core/i18n'
     import type { IDropdownItem } from '@core/utils'
