+++ packages/shared/src/components/RecoveryPhrase.svelte
@@ -1,5 +1,5 @@
 <script lang="ts">
-    import { appSettings } from '@core/app'
+    import { appSettings } from '@core/app/stores'
 
     export let recoveryPhrase: string[] = []
     export let verifyRecoveryPhrase: string[] = undefined
