+++ packages/shared/src/components/buttons/ReceiveAddressButton.svelte
@@ -1,6 +1,6 @@
 <script lang="ts">
     import { selectedAccount } from '@core/account'
-    import { appSettings } from '@core/app'
+    import { appSettings } from '@core/app/stores'
     import { localize } from '@core/i18n'
     import { QR, Text, FontWeight, AddressBox } from '@ui'
