+++ packages/shared/src/lib/core/app/index.ts
@@ -1,7 +1,5 @@
-export * from './actions'
 export * from './constants'
 export * from './classes'
 export * from './enums'
 export * from './interfaces'
-export * from './stores'
 export * from './utils'
